#1832 Moved CompilationRule from the test class into a standalone package "test_utils".
